Add all missing PHPDoc documentation
This patch intentionally does nothing but documenting the current
status quo, i.e. what these methods are actually able to return.
Some of this is really strange, e.g. code uses the return value
of a method that doesn't return anything. There are also some
edge-case code paths that don't end with a return. These look
like mistakes. It's not the point of this patch to fix any of
this, only to document it.

Bug: T260226
Change-Id: I00cce3bfb8da552aca826f72f97b303d48832202